+++ src/api/index.ts
@@ -1,3 +1,4 @@
 export * from "./nodeContents"
 export * from "./transports"
 export * from "./objectstructure"
+export * from "./activate"